requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
coloredlogs==15.0.1
cryptography==46.0.5
ddddocr==1.5.6
deprecation==2.1.0
exceptiongroup==1.3.1
fastapi==0.128.8
flatbuffers==25.12.19
fsspec==2025.10.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.13.0.92
packaging==26.0
pdf2image==1.17.0
pillow==11.3.0
postgrest==2.28.2
propcache==0.4.1
protobuf==6.33.5
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.12.1
pyparsing==3.3.2
PyPDF2==3.0.1
pyroaring==1.0.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
realtime==2.28.2
requests==2.32.5
rich==14.3.3
six==1.17.0
sortedcontainers==2.4.0
starlette==0.49.3
storage3==2.28.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.2
supabase-auth==2.28.2
supabase-functions==2.28.2
sympy==1.14.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0